Feature/importsql (#5)
* Feature/create entity (#4)

* Add Board && Member: 엔티티 기본 구조 생성: 페어프로그래밍

* Update Board: Member Entity 와 관계 설정 : 페어프로그래밍

* Add DummyMember: Member 더미 데이터 추가

* Add import.sql: DML 및 h2 설정 변경
