+++ src/main/java/study/tdd/simpleboard/SimpleBoardApplication.java
@@ -5,7 +5,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class SimpleBoardApplication {
+
     public static void main(String[] args) {
         SpringApplication.run(SimpleBoardApplication.class, args);
     }
+
 }
