+++ src/main/resources/application.yml
@@ -4,8 +4,9 @@ spring:
       enabled: true
       path: /h2-console
       settings:
-        web-allow-others: false
+        web-allow-others: true
         trace: false
+
   datasource:
     url: jdbc:h2:mem:testdb;MODE=MySQL;DB_CLOSE_DELAY=-1
     username: sa
@@ -18,5 +19,5 @@ spring:
     properties:
       hibernate:
         format_sql: true
-        #show_sql: true
-        dialect: org.hibernate.dialect.MySQLDialect
+        show_sql: true
+        dialect: org.hibernate.dialect.MySQL5Dialect
