+++ src/main/resources/import.sql
@@ -1,0 +1,40 @@
+insert into MEMBER (member_email, nickname, password) values ('find1086@mail.com', 'a', '1234');
+insert into MEMBER (member_email, nickname, password) values ('pursue503@mail.com', 'a', '1234');
+insert into MEMBER (member_email, nickname, password) values ('aa@mail.com', 'a', '1234');
+insert into MEMBER (member_email, nickname, password) values ('bb@mail.com', 'b', '1234');
+insert into MEMBER (member_email, nickname, password) values ('cc@mail.com', 'c', '1234');
+insert into MEMBER (member_email, nickname, password) values ('dd@mail.com', 'd', '1234');
+insert into MEMBER (member_email, nickname, password) values ('ee@mail.com', 'e', '1234');
+insert into MEMBER (member_email, nickname, password) values ('ff@mail.com', 'f', '1234');
+insert into MEMBER (member_email, nickname, password) values ('gg@mail.com', 'g', '1234');
+insert into MEMBER (member_email, nickname, password) values ('hh@mail.com', 'h', '1234');
+insert into MEMBER (member_email, nickname, password) values ('ii@mail.com', 'i', '1234');
+insert into MEMBER (member_email, nickname, password) values ('jj@mail.com', 'j', '1234');
+insert into MEMBER (member_email, nickname, password) values ('kk@mail.com', 'k', '1234');
+insert into MEMBER (member_email, nickname, password) values ('mm@mail.com', 'm', '1234');
+insert into MEMBER (member_email, nickname, password) values ('pp@mail.com', 'p', '1234');
+
+insert into BOARD (boardtitle, member_id) values ('1번 게시물 제목입니다.', 1);
+insert into BOARD (boardtitle, member_id) values ('2번 게시물 제목입니다.', 1);
+insert into BOARD (boardtitle, member_id) values ('3번 게시물 제목입니다.', 2);
+insert into BOARD (boardtitle, member_id) values ('4번 게시물 제목입니다.', 2);
+insert into BOARD (boardtitle, member_id) values ('5번 게시물 제목입니다.', 2);
+insert into BOARD (boardtitle, member_id) values ('6번 게시물 제목입니다.', 2);
+insert into BOARD (boardtitle, member_id) values ('7번 게시물 제목입니다.', 3);
+insert into BOARD (boardtitle, member_id) values ('8번 게시물 제목입니다.', 3);
+insert into BOARD (boardtitle, member_id) values ('9번 게시물 제목입니다.', 3);
+insert into BOARD (boardtitle, member_id) values ('10번 게시물 제목입니다.', 4);
+insert into BOARD (boardtitle, member_id) values ('11번 게시물 제목입니다.', 4);
+insert into BOARD (boardtitle, member_id) values ('12번 게시물 제목입니다.', 4);
+insert into BOARD (boardtitle, member_id) values ('13번 게시물 제목입니다.', 5);
+insert into BOARD (boardtitle, member_id) values ('14번 게시물 제목입니다.', 5);
+insert into BOARD (boardtitle, member_id) values ('15번 게시물 제목입니다.', 7);
+insert into BOARD (boardtitle, member_id) values ('16번 게시물 제목입니다.', 7);
+insert into BOARD (boardtitle, member_id) values ('17번 게시물 제목입니다.', 10);
+insert into BOARD (boardtitle, member_id) values ('18번 게시물 제목입니다.', 10);
+insert into BOARD (boardtitle, member_id) values ('19번 게시물 제목입니다.', 10);
+insert into BOARD (boardtitle, member_id) values ('20번 게시물 제목입니다.', 12);
+insert into BOARD (boardtitle, member_id) values ('21번 게시물 제목입니다.', 12);
+insert into BOARD (boardtitle, member_id) values ('22번 게시물 제목입니다.', 12);
+insert into BOARD (boardtitle, member_id) values ('23번 게시물 제목입니다.', 12);
+insert into BOARD (boardtitle, member_id) values ('24번 게시물 제목입니다.', 12);
